% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_results}
\alias{compute_results}
\title{Compute the variable inclusion probability}
\usage{
compute_results(results)
}
\arguments{
\item{results}{list containing coefficients of glmnet model from each bootstrapped run}
}
\value{
a data.frame containg the variable name and its probability of inclusion
}
\description{
Computes the variable inclusion probability from bootstrapped lasso runs.
}
\details{
Computes the variable inclusion probability as the fraction of non-zero coefficients 
out of all bootstrapped lasso runs.
}
