% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggleSample-data.R
\docType{data}
\name{kaggleSample}
\alias{kaggleSample}
\title{Random Sample from Santander Customer Transaction data set from kaggle competition.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 2500 rows and 202 columns.}
\source{
https://kaggle.com/c/santander-customer-transaction-prediction/data
}
\usage{
data(kaggleSample)
}
\description{
Data was included as it provides good example. N = 2500, p = 200, response balance: o(0) = 2245, o(1) = 255
}
\examples{
data(kaggleSample)
\donttest{
# heat map of correlation matrix, linked to scatterplots
iplotCorr(geneExpr$expr, geneExpr$genotype, reorder=TRUE)}
}
\keyword{datasets}
