% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_indices}
\alias{sample_indices}
\title{Sample column names and row indices}
\usage{
sample_indices(feature_columns, nrows_, column_proportion, record_proportion)
}
\arguments{
\item{feature_columns}{character vector containing names to sample from}

\item{nrows_}{the number of rows}

\item{column_proportion}{numeric specifying the fraction of columns to include in the sample}

\item{record_proportion}{numeric specifying the fraction of rows in include in the sample}
}
\value{
list
\describe{
 \item{random_rows}{Vector containing row indices}
 \item{random_columns}{Vector containing feature names}
}
}
\description{
Sample column names and row indices
}
\seealso{
Other sampling: 
\code{\link{sample_matrix}()}
}
\concept{sampling}
