% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_matrix}
\alias{sample_matrix}
\title{Sample rows and columns of a matrix or data.frame}
\usage{
sample_matrix(
  df,
  target_column,
  feature_columns,
  column_proportion,
  record_proportion
)
}
\arguments{
\item{df}{data.frame}

\item{target_column}{str}

\item{feature_columns}{character vector containing names to sample from}

\item{column_proportion}{numeric specifying the fraction of columns to include in the sample}

\item{record_proportion}{numeric specifying the fraction of rows in include in the sample}
}
\value{
list
\describe{
 \item{x}{Subset of original data.frame containing sampled rows and features}
 \item{y}{vector containing target values corresponding to the rows of x}
}
}
\description{
Sample rows and columns of a matrix or data.frame
}
\seealso{
Other sampling: 
\code{\link{sample_indices}()}
}
\concept{sampling}
