% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoSelector.R
\name{lassoSelector}
\alias{lassoSelector}
\title{L1 Lasso Sellection From Dense Predictor Space.}
\usage{
lassoSelector(df, target_column, feature_columns, column_proportion,
  n_iterations, l1_lambda, glmnet_family)
}
\arguments{
\item{df}{\code{data.table}-type data matrix}

\item{target_column}{Character string identifying the target column}

\item{feature_columns}{Character vector identifying the feature columns to investigate}

\item{column_proportion}{Numeric on the interval (0,1), this is the percentage of columns randomly sampeled}

\item{n_iterations}{The number of bootstrap replications}

\item{l1_lambda}{The LASSO L1 penalty parameter}

\item{glmnet_family}{Character string, see glmnet for more information}
}
\value{
data.table object of Conditional Variable Inclusion Probability
}
\description{
L1 Lasso Sellection From Dense Predictor Space.
}
